+++ jodag-spring/src/test/java/jodag/entity/hibernate/EmbeddableClass.java
@@ -1,4 +1,4 @@
-package jodag.hibernate;
+package jodag.entity.hibernate;
 
 import jakarta.persistence.Embeddable;
