+++ jodag-spring/src/test/java/jodag/entity/hibernate/MemberType.java
@@ -1,4 +1,4 @@
-package jodag.hibernate;
+package jodag.entity.hibernate;
 
 public enum MemberType {
     A, B, C, D;
