+++ jodag-spring/src/test/java/jodag/entity/mybatis/associations/GrandParent.java
@@ -1,4 +1,4 @@
-package jodag.mybatis.associations;
+package jodag.entity.mybatis.associations;
 
 import java.util.List;
