+++ jodag-spring/src/test/java/jodag/entity/mybatis/associations/GrandParentMapper.java
@@ -1,4 +1,4 @@
-package jodag.mybatis.associations;
+package jodag.entity.mybatis.associations;
 
 import org.apache.ibatis.annotations.Mapper;
