+++ jodag-spring/src/test/java/jodag/entity/mybatis/associations/Parent.java
@@ -1,38 +1,19 @@
-package jodag.hibernate.associations;
+package jodag.entity.mybatis.associations;
 
-
-import jakarta.persistence.*;
 import jodag.annotation.ValueSource;
 
 import java.util.List;
 
-@Entity
 public class Parent {
-    @Id
+
     private Long id;
 
-    @ValueSource(generatorKey = "email")
+    @ValueSource(generatorKey = "name")
     private String name;
 
-    private String test;
-
-    @OneToMany(mappedBy = "parent", cascade = CascadeType.ALL)
     private List<Child> children;
-
-    @ManyToOne(fetch = FetchType.LAZY)
     private GrandParent grandParent;
 
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getTest() {
-        return test;
-    }
 
     public List<Child> getChildren() {
         return children;
@@ -47,6 +28,7 @@ public class Parent {
         return "Parent{" +
                 "id=" + id +
                 ", name='" + name + '\'' +
+                ", children=" + children +
                 '}';
     }
 }
