+++ jodag-spring/src/test/java/jodag/generatorTest/EntityInstanceCreatorTest.java
@@ -3,8 +3,8 @@ package jodag.generatorTest;
 import jodag.JodagTestApplication;
 import jodag.generator.Generator;
 import jodag.generator.SpringGeneratorFactory;
-import jodag.hibernate.EmbeddableClass;
-import jodag.hibernate.TestEntity;
+import jodag.entity.hibernate.EmbeddableClass;
+import jodag.entity.hibernate.TestEntity;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
