+++ jodag-spring/src/test/java/jodag/mybatisTest/GenerateTypeChildTest.java
@@ -4,9 +4,9 @@ import jodag.JodagTestApplication;
 import jodag.generator.EntityGenerator;
 import jodag.generator.GenerateType;
 import jodag.generator.SpringGeneratorFactory;
-import jodag.mybatis.associations.Child;
-import jodag.mybatis.associations.GrandParent;
-import jodag.mybatis.associations.Parent;
+import jodag.entity.mybatis.associations.Child;
+import jodag.entity.mybatis.associations.GrandParent;
+import jodag.entity.mybatis.associations.Parent;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
