+++ jodag-spring/src/test/resources/mapper/ChildMapper.xml
@@ -1,9 +1,9 @@
 <!-- resources/mapper/ChildMapper.xml -->
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="jodag.mybatis.associations.ChildMapper">
+<mapper namespace="jodag.entity.mybatis.associations.ChildMapper">
 
-    <resultMap id="ChildResultMap" type="jodag.mybatis.associations.Child">
+    <resultMap id="ChildResultMap" type="jodag.entity.mybatis.associations.Child">
         <id property="id" column="c_id"/>
         <result property="name" column="c_name"/>
         <association property="parent" resultMap="ParentResultMap"/>
