+++ jodag-spring/src/test/resources/mapper/GrandParentMapper.xml
@@ -1,9 +1,9 @@
 <!-- resources/mapper/GrandparentMapper.xml -->
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="jodag.mybatis.associations.GrandParentMapper">
+<mapper namespace="jodag.entity.mybatis.associations.GrandParentMapper">
 
-    <resultMap id="GrandParentResultMap" type="jodag.mybatis.associations.GrandParent">
+    <resultMap id="GrandParentResultMap" type="jodag.entity.mybatis.associations.GrandParent">
         <id property="id" column="gp_id"/>
         <result property="name" column="gp_name"/>
         <collection property="parents" resultMap="ParentResultMap"/>
