+++ jodag-spring/src/test/resources/mapper/ParentMapper.xml
@@ -1,9 +1,9 @@
 <!-- resources/mapper/ParentMapper.xml -->
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="jodag.mybatis.associations.ParentMapper">
+<mapper namespace="jodag.entity.mybatis.associations.ParentMapper">
 
-    <resultMap id="ParentResultMap" type="jodag.mybatis.associations.Parent">
+    <resultMap id="ParentResultMap" type="jodag.entity.mybatis.associations.Parent">
         <id property="id" column="p_id"/>
         <result property="name" column="p_name"/>
         <association property="grandParent"  resultMap="GrandParentResultMap"/>
